@@ -1,27 +1,30 @@
 {
   "name": "object-inspect",
-  "version": "1.12.0",
+  "version": "1.12.1",
   "description": "string representations of objects in node and the browser",
   "main": "index.js",
   "sideEffects": false,
   "devDependencies": {
-    "@ljharb/eslint-config": "^20.1.0",
-    "aud": "^1.1.5",
-    "auto-changelog": "^2.3.0",
+    "@ljharb/eslint-config": "^21.0.0",
+    "aud": "^2.0.0",
+    "auto-changelog": "^2.4.0",
     "core-js": "^2.6.12",
-    "error-cause": "^1.0.3",
-    "es-value-fixtures": "^1.2.1",
-    "eslint": "^8.5.0",
+    "error-cause": "^1.0.4",
+    "es-value-fixtures": "^1.4.1",
+    "eslint": "=8.8.0",
     "for-each": "^0.3.3",
-    "functions-have-names": "^1.2.2",
+    "functions-have-names": "^1.2.3",
     "has-tostringtag": "^1.0.0",
     "make-arrow-function": "^1.2.0",
+    "mock-property": "^1.0.0",
+    "npmignore": "^0.3.0",
     "nyc": "^10.3.2",
     "safe-publish-latest": "^2.0.0",
     "string.prototype.repeat": "^1.0.0",
-    "tape": "^5.3.2"
+    "tape": "^5.5.3"
   },
   "scripts": {
+    "prepack": "npmignore --auto --commentLines=autogenerated",
     "prepublish": "not-in-publish || npm run prepublishOnly",
     "prepublishOnly": "safe-publish-latest",
     "pretest": "npm run lint",
@@ -80,5 +83,11 @@
     "commitLimit": false,
     "backfillLimit": false,
     "hideCredit": true
+  },
+  "publishConfig": {
+    "ignore": [
+      ".github/workflows",
+      "./test-core-js.js"
+    ]
   }
 }
