@@ -4,23 +4,22 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import "bootstrap/dist/css/bootstrap.min.css";
-import { Provider} from 'react-redux'
+import { Provider} from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 import persistStore from 'redux-persist/es/persistStore';
 import store from "./store";
 
 const persistedStore = persistStore(store);
-//const root = ReactDOM.createRoot(document.getElementById('root'));
-ReactDOM.render(
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-  <Provider store={store}>
-  <PersistGate loading={null} persistor={persistedStore}>
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistedStore}>
     <App />
     </PersistGate>
     </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
-
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
