@@ -1,6 +1,6 @@
- const mongoose = require('mongoose');
- require('dotenv').config();
+const mongoose = require('mongoose');
+require('dotenv').config();
 
- mongoose.connect(`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PW}@cluster0.wyptw.mongodb.net/Chat?retryWrites=true&w=majority`,()=> {
-     console.log('connected to mongodb')
- })
+mongoose.connect(`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PW}@cluster0.wyptw.mongodb.net/Nimbuzz?retryWrites=true&w=majority`,()=> {
+    console.log('connected to mongodb')
+})
