@@ -1,23 +1,23 @@
-import { createSlice } from "@reduxjs/toolkit";
+import{ createSlice } from "@reduxjs/toolkit";
 import appApi from "../services/appApi";
 
 export const userSlice = createSlice({
     name: "user",
     initialState: null,
     reducers: {
-        addNotifications: (state,{ payload}) => {},
-        resetNotifications: (state,{ payload}) => {},
+        addNotifications: (state, { payload }) => {},
+        resetNotifications: (state, { payload }) => {},
     },
 
     extraReducers: (builder) => {
-        //save user after signup
-        builder.addMatcher(appApi.endpoints.signupUser.matchFulfilled, (state,{ payload}) => payload);
-        //save user after login
-        builder.addMatcher(appApi.endpoints.loginUser.matchFulfilled, (state,{ payload}) => payload);
-         // logout: destory user session
-         builder.addMatcher(appApi.endpoints.signupUser.matchFulfilled, () => null );
-        },
+        // save user after signup
+        builder.addMatcher(appApi.endpoints.signupUser.matchFulfilled, (state, {payload}) => payload);
+        // save user after login
+        builder.addMatcher(appApi.endpoints.loginUser.matchFulfilled, (state, {payload}) => payload);
+        // logout: destroy user session
+        builder.addMatcher(appApi.endpoints.logoutUser.matchFulfilled, (state, {payload}) => payload);
+    },
 });
 
 export const { addNotifications, resetNotifications } = userSlice.actions;
-export default userSlice.reducer; 
+export default userSlice.reducer;
