@@ -1,15 +1,15 @@
-import { createApi, fetchBaseQuery} from '@reduxjs/toolkit/query/react';
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
-//define a service user a base Url
+// define a service user a base URL
 
 const appApi = createApi({
-    reducePath: 'appApi',
+    reducerPath: 'appApi',
     baseQuery: fetchBaseQuery({
-        baseUrl: 'http://localhost:5001'
+        baseUrl: 'http://localhost:5001',
     }),
 
     endpoints: (builder) => ({
-        //creating the user
+        // creating the user
         signupUser: builder.mutation({
             query: (user) => ({
                 url: "/users",
@@ -18,12 +18,12 @@ const appApi = createApi({
             }),
         }),
 
-        //login
+        // login
         loginUser: builder.mutation({
             query: (user) => ({
-            url: "/users/login",
-            method: "POST",
-            body: user,
+                url: "/users/login",
+                method: "POST",
+                body: user,
             }),
         }),
 
@@ -33,14 +33,11 @@ const appApi = createApi({
                 url: "/logout",
                 method: "DELETE",
                 body: payload,
-                }),
+            }),
         }),
-
-
-
     }),
 });
 
-export const {userSignupUserMutation, useLoginUserMutation, useLogoutUserMutation  } = appApi;
+export const { useSignupUserMutation, useLoginUserMutation, useLogoutUserMutation } = appApi;
 
 export default appApi;
