@@ -1,33 +1,34 @@
 import { configureStore } from "@reduxjs/toolkit";
-import userSlice from "./features/userSlice";
+import userSlice from "./features/userSlice"; 
 import appApi from "./services/appApi";
 
-//persist our store
+// persist our store
 import storage from "redux-persist/lib/storage";
-import { combineReducers  } from "redux";
+import { combineReducers } from "redux";
 import { persistReducer } from "redux-persist";
 import thunk from "redux-thunk";
 
-//reducers
+// reducers
 const reducer = combineReducers({
-   user: userSlice,
-   [appApi.reducerPath] : appApi.reducer, 
+    user: userSlice,
+    [appApi.reducerPath]: appApi.reducer,
 });
 
 const persistConfig = {
-    key: 'root',
+    key: "root",
     storage,
     blackList: [appApi.reducerPath],
 };
 
-//persist our store
+// persist our store
 
 const persistedReducer = persistReducer(persistConfig, reducer);
 
-//creating the store
+// creating the store
 
 const store = configureStore({
     reducer: persistedReducer,
     middleware: [thunk, appApi.middleware],
 });
+
 export default store;
